requirements.txt
discord
python-dotenv
beautifulsoup4
scikit-learn
joblib
pandas
numpy